requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
click==8.1.7
colorama==0.4.6
cssbeautifier==1.14.9
Django==4.2.5
django-environ==0.11.2
django-sass==1.1.0
djangorestframework==3.14.0
djlint==1.33.0
EditorConfig==0.12.3
fontawesomefree==6.4.2
html-tag-names==0.1.2
html-void-elements==0.1.0
jsbeautifier==1.14.9
json5==0.9.14
libsass==0.22.0
pathspec==0.11.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.1
typing==3.7.4.3
typing-extensions==4.8.0
